x/sale/types/message_cancel_nft_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"swap/testutil/sample"
)

func TestMsgCancelNFT_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCancelNFT
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCancelNFT{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCancelNFT{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
